+++ gd-bbpress-attachments/code/tools.php
@@ -23,7 +23,7 @@ class GDATTTools {
 		global $wpdb;
 
 		$sql = "SELECT `post_id`, COUNT(*) AS `items` FROM $wpdb->postmeta WHERE `meta_key` = '_bbp_attachment_upload_error' GROUP BY `post_id`";
-		$raw = $wpdb->get_results( $sql, ARRAY_A );
+		$raw = $wpdb->get_results( $sql, ARRAY_A ); // phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared,WordPress.DB.DirectDatabaseQuery.DirectQuery,WordPress.DB.DirectDatabaseQuery.NoCaching
 
 		return array(
 			'list'   => wp_list_pluck( $raw, 'items', 'post_id' ),
@@ -37,12 +37,11 @@ class GDATTTools {
 	public function delete_all_logged_errors() {
 		global $wpdb;
 
-		$sql = "DELETE FROM $wpdb->postmeta WHERE `meta_key` = '_bbp_attachment_upload_error'";
-		$wpdb->query( $sql );
+		$wpdb->query( "DELETE FROM $wpdb->postmeta WHERE `meta_key` = '_bbp_attachment_upload_error'" ); //phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery,WordPress.DB.DirectDatabaseQuery.NoCaching
 	}
 
 	public function process_action() {
-		$nonce = isset( $_GET['_wpnonce'] ) ? sanitize_text_field( $_GET['_wpnonce'] ) : '';
+		$nonce = isset( $_GET['_wpnonce'] ) ? sanitize_text_field( wp_unslash( $_GET['_wpnonce'] ) ) : '';
 		$url   = remove_query_arg( array( '_wpnonce', 'action' ) );
 
 		if ( wp_verify_nonce( $nonce, 'gdatt-clear-error-log' ) ) {